Refactor blend tests to gtest
