+++ changes/6059.change.md
@@ -1,0 +1,1 @@
+Refactored `blend` to be callable and migrated its legacy Makefile tests to gtest/CTest. Replaced the legacy ISISTESTDATA dependency with reduced repo-local fixtures, reducing blend test data footprint from 118 MB to 12 MB.
