+++ isis/src/base/apps/blend/blend.h
@@ -1,0 +1,10 @@
+#ifndef blend_h
+#define blend_h
+
+#include "UserInterface.h"
+
+namespace Isis {
+  extern void blend(UserInterface &ui);
+}
+
+#endif
