--- isis/src/base/apps/blend/tsts/Makefile
@@ -1,4 +1,0 @@
-BLANKS = "%-6s"    
-LENGTH = "%-40s"
-
-include $(ISISROOT)/make/isismake.tststree
